requirements.txt
APScheduler==3.6.3
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
decorator==5.0.7
defusedxml==0.7.1
Django==2.2.5
django-allauth==0.39.1
django-cors-headers==3.7.0
django-debug-toolbar==3.2.1
django-environ==0.4.5
django-extensions==2.1.6
django-filter==2.1.0
django-rest-swagger==2.2.0
djangorestframework==3.9.3
drf-yasg==1.17.1
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
MarkupSafe==1.1.1
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==20.9
Pillow==8.2.0
pycparser==2.20
pyparsing==2.4.7
python-telegram-bot==13.0
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
simplejson==3.17.2
six==1.15.0
sqlparse==0.3.1
tornado==6.1
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.25.10
whitenoise==4.1.2
